pyproject.toml
[project]
name = "porter"
version = "0.1.0"
description = "Riftbound card extractor using OpenAI Vision"
authors = [{ name = "you" }]
requires-python = ">=3.9"

dependencies = [
    "openai>=1.35.8",
    "pillow>=10.0.0",
    "pydantic>=2.4.0",
    "pytest>=8.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0"
]

[project.scripts]
porter = "porter.cli:main"
